Spring data

+++ main/src/main/java/edu/senla/entity/TypeOfContainer.java
@@ -18,7 +18,7 @@ public class TypeOfContainer implements Serializable{
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Column(name = "number_of_calories")
-    private int numberOfCalories;
+    private long numberOfCalories;
 
     private String name;
