+++ main/src/main/java/edu/senla/service/serviceinterface/TypeOfContainerServiceInterface.java
@@ -6,11 +6,11 @@ public interface TypeOfContainerServiceInterface {
 
     public TypeOfContainerDTO createTypeOfContainer(TypeOfContainerDTO newTypeOfContainerDTO);
 
-    public TypeOfContainerDTO readTypeOfContainer(int id);
+    public TypeOfContainerDTO readTypeOfContainer(long id);
 
-    public TypeOfContainerDTO updateTypeOfContainer(int id, TypeOfContainerDTO updatedTypeOfContainerDTO);
+    public TypeOfContainerDTO updateTypeOfContainer(long id, TypeOfContainerDTO updatedTypeOfContainerDTO);
 
-    public void deleteTypeOfContainer(int id);
+    public void deleteTypeOfContainer(long id);
 
     public boolean isTypeOfContainerExists(TypeOfContainerDTO typeOfContainer);
